+++ khoros/structures/boards.py
@@ -6,7 +6,7 @@
 :Example:           ``board_url = boards.create(khoros_object, 'my-board', 'My Board', 'forum', return_url=True)``
 :Created By:        Jeff Shurtliff
 :Last Modified:     Jeff Shurtliff
-:Modified Date:     18 May 2020
+:Modified Date:     23 May 2020
 """
 
 import warnings
@@ -24,7 +24,7 @@ def create(khoros_object, board_id, board_title, discussion_style, description=N
            one_entry_per_contest=None, one_kudo_per_contest=None, posting_date_end=None, posting_date_start=None,
            voting_date_end=None, voting_date_start=None, winner_announced_date=None, full_response=None, return_id=None,
            return_url=None, return_api_url=None, return_http_code=None, return_status=None,
-           return_developer_message=None):
+           return_error_messages=None, split_errors=False):
     """This function creates a new board within a Khoros Community environment.
 
     :param khoros_object: The core :py:class:`khoros.Khoros` object
@@ -97,9 +97,11 @@ def create(khoros_object, board_id, board_title, discussion_style, description=N
     :type return_http_code: bool, None
     :param return_status: Determines if the **Status** of the API response should be returned by the function
     :type return_status: bool, None
-    :param return_developer_message: Determines if the **Developer Response Message** (if any) associated with the
+    :param return_error_messages: Determines if the **Developer Response Message** (if any) associated with the
            API response should be returned by the function
-    :type return_developer_message: bool, None
+    :type return_error_messages: bool, None
+    :param split_errors: Defines whether or not error messages should be merged when applicable
+    :type split_errors: bool
     :returns: Boolean value indicating a successful outcome (default), the full API response or one or more specific
               fields defined by function arguments
     :raises: :py:exc:`khoros.errors.exceptions.MissingRequiredDataError`,
@@ -118,7 +120,7 @@ def create(khoros_object, board_id, board_title, discussion_style, description=N
     headers = {'content-type': 'application/json'}
     response = api.post_request_with_retries(api_url, payload, khoros_object=khoros_object, headers=headers)
     return api.deliver_v2_results(response, full_response, return_id, return_url, return_api_url, return_http_code,
-                                  return_status, return_developer_message)
+                                  return_status, return_error_messages, split_errors)
 
 
 def structure_payload(khoros_object, board_id, board_title, discussion_style, description=None, parent_category_id=None,